+++ python/paddle/distributed/fleet/meta_parallel/__init__.py
@@ -13,5 +13,5 @@
 # limitations under the License.
 
 from .parallel_layers import *
-from .model_parallel import ModelParallel
+from .tensor_parallel import TensorParallel
 from .pipeline_parallel import PipelineParallel
